+++ ChangeLog.md
@@ -3,6 +3,9 @@ XP AST ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Add support for generic wildcards such as `Filter<?>`, resolving
+  ambiguity with nullable types
+  (@thekid)
 * Resolve ambiguity between short open tag and nullables in generics,
   e.g. `Filter<?int>`.
   (@thekid)